moved load_configs.py from scripts to src
