+++ monitoring/refresh_monitoring_dashboard.py
@@ -13,7 +13,7 @@ from evidently.metrics import DriftedColumnsCount, MissingValueCount, ValueDrift
 from tools import load_model_artifacts, prepare_data_for_monitoring
 
 from data import load_data
-from scripts.load_configs import Configs
+from load_configs import Configs
 
 warnings.simplefilter(action="ignore", category=FutureWarning)
 warnings.simplefilter(action="ignore", category=RuntimeWarning)
